Remove Peer-Directed Projects Center, PDPC references
From: Tom Wesley <tom@tomaw.net>
To: freenode-bikeshed@freenode.net
Date: Mon, 17 Dec 2012 10:43:35 +0000
Subject: [freenode-bikeshed] PDPC closure and how this impacts you

Hi,

As you may well have heard by now PDPC will be closing shortly.  This will have
minimal impact on the day to day operations of freenode but it's likely to
provoke some questions from users.

While it's been discussed on IRC repeatedly for a couple of years it's perhaps
useful to give a little background. It's been nearly two years since we sent
HMRC the first application to be struck off which we've finlly had a supportive
acknowledgement.
PDPC itself has been operating at a loss ever since its second year in the UK
(in year 1 we met our targets purely due to the large donation by canonical and
the large donations made by staff members.) Due to the reduction in donations
we do not meet the threshold for charitable status, which was the main
objective for forming the organisation in the first place. Overall it has got
to the point where the financial and administritive overhead outweighs any
advantage.

The infrastructure team is working to relocate the services provided by PDPC
sponsored machines but there should be no noticable impact to staff or users.

Thanks,
--
Tom Wesley <tom@tomaw.net>