Change/replace irc:// URLs with webchat links
Several of the irc:// channel links were broken. Different browsers
always prepend a #, sometimes prepend a #, or don't prepend a # to the
channel name. This makes it impossible to create an irc:// link to a
channel (particularly a ## channel) that won't be broken or go to the
wrong place on some browser(s). Using webchat links still allows the
click-to-join behavior without sending users to the wrong channel, and
doesn't rely on irc:// protocol support.