Remove hyperion page, and "freenode and IRC"
Hyperion is only relevant as a historical curiosity in passing,
anymore.

The freenode and IRC page doesn't say much beyond duplicating
information from the development pages and "freenode is different, and
so is our ircd/services".  Philosophy covers how and why we're
different, and the development pages can cover the technical
implementation and justification.