Update irssi cap_sasl.pl to 'require' not 'use'
As pointed out by grawity in #freenode, the purpose of the eval is to
make the Crypt and Math modules optional, falling back gracefully to
support only for PLAIN if the modules are not available. This only
works if the eval includes the modules with 'require' (evaluated at
run-time) instead of 'use' (evaluated at compile-time).