Add disclaimer to policy about use of the network
The various crypocoins are becoming popular, and many of them appear to
be using freenode as their IRC home.  We don't want users to see
freenode as a critical resource to effect their transactions, nor do we
want to accept any liability should some important transaction fail to
happen because of network instability or whatever.  Try to make this
more clear to users.