Update and consolidate Tor documentation
Move most of the Tor how-to out of policy and onto the IRC servers
page with the rest of the Tor information. Add a description of exit
policy limitations for those who want to be an exit node. Remove the
solicitation for feedback on the secondary addresses. (Sometimes
they help; using a different port on the same hidden address
doesn't.)